+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 mout changelog
 ==============
 
+Next
+----
+
+ - Add optional delimiter to `string/unCamelCase`
+
 v0.5.0 (2013/04/04)
 -------------------